From sibo + remove code duplication
