--- Code/attention/__init__.py
@@ -1,2 +1,0 @@
-from .transformer import LocalFeatureTransformer
-from .position_encoding import PositionEncodingSine
