--- Code/dataset/dtu/lists/temp.txt
@@ -1,1 +1,0 @@
-scan19
