--- Code/dataset/dtu/lists/test.txt
@@ -1,15 +1,0 @@
-scan24
-scan37
-scan40
-scan55
-scan63
-scan65
-scan69
-scan83
-scan97
-scan105
-scan106
-scan110
-scan114
-scan118
-scan122
