--- Code/dataset/dtu/lists/train.txt
@@ -1,75 +1,0 @@
-scan1
-scan4
-scan5
-scan6
-scan8
-scan9
-scan10
-scan11
-scan12
-scan13
-scan14
-scan15
-scan16
-scan17
-scan18
-scan19
-scan20
-scan21
-scan22
-scan23
-scan28
-scan29
-scan30
-scan31
-scan32
-scan33
-scan34
-scan35
-scan36
-scan38
-scan39
-scan41
-scan42
-scan43
-scan44
-scan45
-scan46
-scan47
-scan48
-scan49
-scan50
-scan51
-scan52
-scan59
-scan60
-scan61
-scan62
-scan64
-scan74
-scan75
-scan76
-scan77
-scan84
-scan85
-scan86
-scan87
-scan88
-scan89
-scan90
-scan91
-scan92
-scan93
-scan94
-scan95
-scan96
-scan98
-scan99
-scan100
-scan101
-scan102
-scan103
-scan104
-scan126
-scan127
-scan128
